added a LOT of none checks

+++ main.py
@@ -64,8 +64,6 @@ sentence_tabs = []
 class_students = []
 preset_list = {}
 sentences = []  # Should be populated with SentenceGroup elements
-
-grade_parser = GradeSet()
 grade_rules = []
 
 first_run = True
@@ -290,22 +288,23 @@ def fill_class_data():
     ro = row_offset #Row offset
 
     #Drop the headers, iterate over all student rows
-    for student in current_class[1:]:
-        class_students.append(
-            Student(
-                student[0], #First Name
-                student[1] if len(student) >= 2 else "", #Last Name
-                student[2] if len(student) >= 3 else "", #Gender
-                student[3] if len(student) >= 4 else "", #Report
-                #Zip assignment names to grades
-                dict(zip(current_class[0][4:], student[4:])) if len(student) >= 5 else {},
-                class_dropdown.currentText(),
-                ro
+    if current_class is not None:
+        for student in current_class[1:]:
+            class_students.append(
+                Student(
+                    student[0], #First Name
+                    student[1] if len(student) >= 2 else "", #Last Name
+                    student[2] if len(student) >= 3 else "", #Gender
+                    student[3] if len(student) >= 4 else "", #Report
+                    #Zip assignment names to grades
+                    dict(zip(current_class[0][4:], student[4:])) if len(student) >= 5 else {},
+                    class_dropdown.currentText(),
+                    ro
+                )
             )
-        )
-        ro+=1
+            ro+=1
 
-    student_dropdown.addItems([student.first_name + " " + student.last_name for student in class_students])
+        student_dropdown.addItems([student.first_name + " " + student.last_name for student in class_students])
     if class_dropdown.history[-1][0] != class_dropdown.currentText()[0]:
         update_sentences()
     class_dropdown.history.append(class_dropdown.currentText())
@@ -337,24 +336,26 @@ def update_sentences():
     global sentence_tabs
     global class_dropdown
 
-    for tab in sentence_tabs:
-        if tab[0][-1] == class_dropdown.currentText()[0]:
-            for elem in sentences:
-                elem.delete()
-            sentences = []
-            print("Called update sentences...")
-            current_sentences = get_sheet(report_sheet, "{}!A1:Z1000".format("Sentences " + str(class_dropdown.currentText()[0])), "COLUMNS").get('values')
-            count = 0
-            ro = 0
-            for entry in current_sentences:
-                if len(entry) > 1:
-                    sentences.append(
-                        SentenceGroup("S{}:".format(count+1), 50, 125 + 25 * count, entry[1:], ro)
-                    )
-                    count += 1
-                ro+=1
-            populate_presets(current_sentences)
-            break
+    if len(sentence_tabs) > 0:
+        for tab in sentence_tabs:
+            if tab[0][-1] == class_dropdown.currentText()[0]:
+                for elem in sentences:
+                    elem.delete()
+                sentences = []
+                print("Called update sentences...")
+                current_sentences = get_sheet(report_sheet, "{}!A1:Z1000".format("Sentences " + str(class_dropdown.currentText()[0])), "COLUMNS").get('values')
+                count = 0
+                ro = 0
+                if current_sentences is not None:
+                    for entry in current_sentences:
+                        if len(entry) > 1:
+                            sentences.append(
+                                SentenceGroup("S{}:".format(count+1), 50, 125 + 25 * count, list(filter(None, entry[1:])), ro)
+                            )
+                            count += 1
+                        ro+=1
+                populate_presets(current_sentences)
+                break
     update_tab_order()
 
 class Preset:
@@ -380,36 +381,36 @@ def populate_presets(sentence_set):
     preset_dropdown.clear()
     presets_found = []
     preset_list = {}
-
     grade_rules = []
 
     count = 1
-    for elem in sentence_set:
-        for cell in elem:
-            #Split cell based on Preset delimiters
-            cell_parts = re.split(f"[{Preset.prefix['grade']}{Preset.prefix['linear']}]", cell)
-            if len(cell_parts) > 1:
-                preset_text = cell_parts[0] #Sentence comes first
-                for ps in cell_parts[1:]:
-                    preset_parts = re.split("[|]", ps) #Split cell based into its components
-                    if len(preset_parts) == 3: #3-part ones have a ruleset string
-                        grade_rules.append(
-                            Preset(preset_text, preset_parts[0], count, ruleset=preset_parts[1], priority=preset_parts[2])
-                        )
-                    elif len(preset_parts) == 2:
-                        presets_found.append(
-                            Preset(preset_text, preset_parts[0], preset_parts[1])
-                        )
-        count+=1
+    if sentence_set is not None:
+        for elem in sentence_set:
+            for cell in elem:
+                #Split cell based on Preset delimiters
+                cell_parts = re.split(f"[{Preset.prefix['grade']}{Preset.prefix['linear']}]", cell)
+                if len(cell_parts) > 1:
+                    preset_text = cell_parts[0] #Sentence comes first
+                    for ps in cell_parts[1:]:
+                        preset_parts = re.split("[|]", ps) #Split cell based into its components
+                        if len(preset_parts) == 3: #3-part ones have a ruleset string
+                            grade_rules.append(
+                                Preset(preset_text, preset_parts[0], count, ruleset=preset_parts[1], priority=preset_parts[2])
+                            )
+                        elif len(preset_parts) == 2:
+                            presets_found.append(
+                                Preset(preset_text, preset_parts[0], preset_parts[1])
+                            )
+            count+=1
 
-    for elem in presets_found:
-        if not elem.group in preset_list:
-            preset_list[elem.group] = []
-        preset_list[elem.group].append(elem)
+        for elem in presets_found:
+            if not elem.group in preset_list:
+                preset_list[elem.group] = []
+            preset_list[elem.group].append(elem)
 
-    for elem in preset_list:
-        preset_list[elem] = sorted(preset_list[elem], key=lambda e: int(e.index))
-        preset_dropdown.addItem(elem)
+        for elem in preset_list:
+            preset_list[elem] = sorted(preset_list[elem], key=lambda e: int(e.index))
+            preset_dropdown.addItem(elem)
 
 
 
@@ -418,18 +419,21 @@ def update_report():
     global student_dropdown
     global report_area
 
-    if student_dropdown is not None:
+    if student_dropdown.count() > 0:
         report_area.setText(class_students[student_dropdown.currentIndex()].report)
-
+    else:
+        report_area.setText("")
+    report_area.repaint()
 
 
 def send_report():
     global class_students
     global student_dropdown
 
-    class_students[student_dropdown.currentIndex()].report = report_area.toPlainText()
-    submit_thread = Thread(target=class_students[student_dropdown.currentIndex()].submit_report)
-    submit_thread.start()
+    if len(class_students) > 0:
+        class_students[student_dropdown.currentIndex()].report = report_area.toPlainText()
+        submit_thread = Thread(target=class_students[student_dropdown.currentIndex()].submit_report)
+        submit_thread.start()
 
 def generate_report():
     global sentences
@@ -438,9 +442,10 @@ def generate_report():
     global class_students
 
     report_area.setText("")
-    for sentence in sentences:
-        if sentence.checkbox.isChecked():
-            report_area.setText(report_area.toPlainText() + replace_generics(sentence.dropdown.currentText())+" ")
+    if len(class_students) > 0:
+        for sentence in sentences:
+            if sentence.checkbox.isChecked():
+                report_area.setText(report_area.toPlainText() + replace_generics(sentence.dropdown.currentText())+" ")
     report_area.repaint()
 
 def generate_report_from_preset():
@@ -461,21 +466,22 @@ def generate_report_from_grades():
     global grade_rules
     global class_dropdown
 
-    current_student = class_students[class_dropdown.currentIndex()]
     report_area.setText("")
-    chosen_options = {}
+    if len(class_students) > 0:
+        current_student = class_students[class_dropdown.currentIndex()]
+        chosen_options = {}
 
-    for elem in grade_rules:
-        matches = lambda G: eval(elem.ruleset, {"__builtins__": None}) #this is a mistake waiting to happen
-        if elem.group in current_student.grades and matches(int(current_student.grades[elem.group])):
-            if elem.index in chosen_options:
-                if chosen_options[elem.index].priority < elem.priority:
+        for elem in grade_rules:
+            matches = lambda G: eval(elem.ruleset, {"__builtins__": None}) #this is a mistake waiting to happen
+            if elem.group in current_student.grades and matches(int(current_student.grades[elem.group])):
+                if elem.index in chosen_options:
+                    if chosen_options[elem.index].priority < elem.priority:
+                        chosen_options[elem.index] = elem
+                else:
                     chosen_options[elem.index] = elem
-            else:
-                chosen_options[elem.index] = elem
 
-    for option in chosen_options:
-        report_area.setText(report_area.toPlainText() + replace_generics(chosen_options[option].text) + " ")
+        for option in chosen_options:
+            report_area.setText(report_area.toPlainText() + replace_generics(chosen_options[option].text) + " ")
     report_area.repaint()
 
 
@@ -494,30 +500,32 @@ def replace_generics(fmt):
     global class_students
     global pronouns
 
-    current_student = class_students[student_dropdown.currentIndex()]
-    ps = current_student.get_pronouns()
+    if len(class_students) > 0:
+        current_student = class_students[student_dropdown.currentIndex()]
+        ps = current_student.get_pronouns()
 
-    #* = Preset
-    #~ = Grade Preset
-    fmt = fmt.split(Preset.prefix['grade'])[0].split(Preset.prefix['linear'])[0].strip().format(
-        name=current_student.first_name,
-        p1=ps[0],
-        p2=ps[1],
-        p3=ps[2],
-        p4=ps[3]
-    ).replace("@", current_student.first_name).replace("#", ps[0]).replace("$", ps[1]).replace("%", ps[2]).replace("^", ps[3])
-    if current_student.gender == "T":
-        fmt = fmt.replace("they is", "they are")
+        #* = Preset
+        #~ = Grade Preset
+        fmt = fmt.split(Preset.prefix['grade'])[0].split(Preset.prefix['linear'])[0].strip().format(
+            name=current_student.first_name,
+            p1=ps[0],
+            p2=ps[1],
+            p3=ps[2],
+            p4=ps[3]
+        ).replace("@", current_student.first_name).replace("#", ps[0]).replace("$", ps[1]).replace("%", ps[2]).replace("^", ps[3])
+        if current_student.gender == "T":
+            fmt = fmt.replace("they is", "they are")
 
-    capitalizationIndices = []
-    for i in range(0, len(fmt)):
-        if fmt[i] == "." or fmt[i] == "!" or fmt[i] == "?": capitalizationIndices.append(i)
+        capitalizationIndices = []
+        for i in range(0, len(fmt)):
+            if fmt[i] == "." or fmt[i] == "!" or fmt[i] == "?": capitalizationIndices.append(i)
 
-    for index in capitalizationIndices:
-        if index + 2 < len(fmt):
-            fmt = fmt[0:index + 2] + fmt[index + 2].upper() + fmt[index + 3:]
+        for index in capitalizationIndices:
+            if index + 2 < len(fmt):
+                fmt = fmt[0:index + 2] + fmt[index + 2].upper() + fmt[index + 3:]
 
-    return fmt.strip()
+        return fmt.strip()
+    else: return None
 
 if __name__ == "__main__":
     app.exec()
